requirements.txt
piper-tts>=1.2.0
soundfile>=0.12.0
langdetect>=1.0.9
numpy>=1.24.0


